Patch structure changes, removes sensitive system
-=# The First Half #=-
When CFX was originally created in 2022, I tried to store patches in
"group" classes named after the mechanic/component they affected to make
storage look cleaner. However, during development I was forced to switch
gears and make future patch mixin classes separate due to a weird issue
where Mixin started refusing to accept inner class patches. While I did
switch gears, up until now the way patches were stored were an
inconsistent mess that made it annoying to find which patches are
located in which projects when working on the project. This half of the
commit migrates ALL patches to just be their own dedicated classes in
me.videogamesm12.cfx.v<version>.patches instead of being a little of all
three columns.

-=# Part 2 #=-
Commit 8f8fdff05c2666e3f3587d8cd13dd32aeb71781d saw the introduction of
the ClickableCommands patch and with it a new means of handling special
patches called "sensitive patches", which were patches that would cause
problems if the regular PatchManager tried to process them
(specifically, the game would crash because classes like Text would be
"loaded too early"), which was because it would load the patch classes
(and by extension, the classes they were trying to hook into) really
early on while tring to get the metadata related to them.

This system was rendered obsolete because I discovered the root cause
of the problem and figured out a better solution, which was implemented
in commit cedfbbc34ac3de4945e095eaad6051605a7cb3ee. This commit removes
the now-obsolete system and migrates its patches to the main system.